fixed the bugs

+++ Debugging_Challenge/TrafficControl/traffic_light_problem.h
@@ -39,7 +39,7 @@ struct lane_of_cars_s
 	int8_t carsThatHaveLeft;
 	
 	int8_t popularity;
-	int8_t timeWaiting;
+	int16_t timeWaiting;
 };
 
 struct intersection_s
